refactored mock request logger to fit current codebase
