+++ coderd/httpmw/loggermock/loggermock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination=mocks/mock_request_logger.go -package=mocks . RequestLogger
+//	mockgen -destination=loggermock/loggermock.go -package=loggermock . RequestLogger
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package loggermock is a generated GoMock package.
+package loggermock
 
 import (
 	context "context"
